+++ wsmaster/codenvy-hosted-api-organization/src/main/java/com/codenvy/organization/api/OrganizationApiModule.java
@@ -18,9 +18,10 @@ import com.codenvy.api.permission.server.SystemDomain;
 import com.codenvy.organization.api.permissions.OrganizationPermissionsProvider;
 import com.codenvy.organization.api.permissions.OrganizationPermissionsFilter;
 import com.codenvy.organization.api.permissions.OrganizationResourceDistributionServicePermissionsFilter;
-import com.codenvy.organization.api.permissions.PersonalOrganizationOwnerPermissionsFilter;
-import com.codenvy.organization.api.permissions.PersonalOrganizationPermissionsFilter;
-import com.codenvy.organization.api.resource.DefaultOrganizationResourcesProvider;
+import com.codenvy.organization.api.personal.PersonalOrganizationOwnerPermissionsFilter;
+import com.codenvy.organization.api.personal.PersonalOrganizationPermissionsFilter;
+import com.codenvy.organization.api.personal.DefaultOrganizationResourcesProvider;
+import com.codenvy.organization.api.personal.PersonalOrganizationProvider;
 import com.codenvy.organization.api.resource.OrganizationResourceLockKeyProvider;
 import com.codenvy.organization.api.resource.OrganizationResourcesDistributionService;
 import com.codenvy.organization.api.resource.OrganizationResourcesPermissionsChecker;
