+++ wsmaster/codenvy-hosted-api-organization/src/main/java/com/codenvy/organization/api/personal/DefaultOrganizationResourcesProvider.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.organization.api.resource;
+package com.codenvy.organization.api.personal;
 
 import com.codenvy.organization.api.OrganizationManager;
 import com.codenvy.organization.shared.model.Organization;
