+++ wsmaster/codenvy-hosted-api-organization/src/main/java/com/codenvy/organization/api/personal/PersonalOrganizationPermissionsFilter.java
@@ -12,17 +12,14 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.organization.api.permissions;
+package com.codenvy.organization.api.personal;
 
-import com.codenvy.api.permission.shared.dto.PermissionsDto;
 import com.codenvy.organization.api.OrganizationManager;
-import com.codenvy.organization.api.permissions.OrganizationDomain;
 import com.codenvy.organization.shared.model.Organization;
 
 import org.eclipse.che.api.core.ForbiddenException;
 import org.eclipse.che.api.core.NotFoundException;
 import org.eclipse.che.api.core.ServerException;
-import org.eclipse.che.api.core.model.user.User;
 import org.eclipse.che.api.user.server.UserManager;
 import org.eclipse.che.everrest.CheMethodInvokerFilter;
 import org.everrest.core.Filter;
@@ -32,19 +29,22 @@ import javax.inject.Inject;
 import javax.ws.rs.Path;
 
 /**
- * Restricts access to setting permissions for owner of organization
+ * Restricts access to remove/rename personal organization
  *
  * @author Sergii Leschenko
  */
 @Filter
-@Path("/permissions/{path:(/.*)?}")
-public class PersonalOrganizationOwnerPermissionsFilter extends CheMethodInvokerFilter {
+@Path("/organization{path:(?!/resource)(/.*)?}")
+public class PersonalOrganizationPermissionsFilter extends CheMethodInvokerFilter {
+    static final String UPDATE_METHOD = "update";
+    static final String REMOVE_METHOD = "remove";
+
     private final OrganizationManager organizationManager;
     private final UserManager         userManager;
 
     @Inject
-    public PersonalOrganizationOwnerPermissionsFilter(OrganizationManager organizationManager,
-                                                      UserManager userManager) {
+    public PersonalOrganizationPermissionsFilter(OrganizationManager organizationManager,
+                                                 UserManager userManager) {
         this.organizationManager = organizationManager;
         this.userManager = userManager;
     }
@@ -54,37 +54,19 @@ public class PersonalOrganizationOwnerPermissionsFilter extends CheMethodInvoker
                                                                                                NotFoundException,
                                                                                                ServerException {
         final String methodName = genericResourceMethod.getMethod().getName();
-
-        final String domainId;
-        final String instanceId;
-        final String userId;
-
         switch (methodName) {
-            case "storePermissions":
-                final PermissionsDto permissions = (PermissionsDto)arguments[0];
-                domainId = permissions.getDomainId();
-                userId = permissions.getUserId();
-                instanceId = organizationManager.getById(permissions.getInstanceId()).getId();
-                break;
-            case "removePermissions":
-                domainId = ((String)arguments[0]);
-                instanceId = ((String)arguments[1]);
-                userId = ((String)arguments[2]);
+            case UPDATE_METHOD:
+            case REMOVE_METHOD:
+                Organization organization = organizationManager.getById((String)arguments[0]);
+                try {
+                    userManager.getByName(organization.getName());
+                    throw new ForbiddenException("It is not allow to rename/remove personal organization.");
+                } catch (NotFoundException ignored) {
+                    // requested organization is not personal
+                }
                 break;
             default:
-                return;
-        }
-
-        if (!OrganizationDomain.DOMAIN_ID.equals(domainId)) {
-            //filter only setting permissions on organization level
-            return;
-        }
-
-        User user = userManager.getById(userId);
-        Organization organization = organizationManager.getById(instanceId);
-
-        if (organization.getName().equals(user.getName())) {
-            throw new ForbiddenException("It is not allowed to edit organization owner's permissions");
+                //do nothing
         }
     }
 }
