+++ wsmaster/codenvy-hosted-api-organization/src/main/java/com/codenvy/organization/api/personal/PersonalOrganizationProvider.java
@@ -12,8 +12,9 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.organization.api;
+package com.codenvy.organization.api.personal;
 
+import com.codenvy.organization.api.OrganizationManager;
 import com.codenvy.organization.api.permissions.OrganizationDomain;
 import com.codenvy.organization.shared.model.Organization;
 import com.codenvy.organization.spi.MemberDao;
