+++ wsmaster/codenvy-sql-schema/src/main/resources/codenvy-schema/5.1.0/1.1__create_personal_organization.sql
@@ -13,6 +13,9 @@
 -- from Codenvy S.A..
 --
 
+-- admin@codenvy.onprem
+-- root-prod
+
 -- Rename existing account to make free old account's names
 UPDATE Account
 SET name=CONCAT('tomigrate', name)
